+++ crates/gen/src/parser/type_reader.rs
@@ -121,7 +121,6 @@ impl TypeReader {
         let exclude = &[
             ("Windows.Foundation", "HResult"),
             ("Windows.Win32.Com", "IUnknown"),
-            ("Windows.Win32.SystemServices", "BOOL"),
             ("Windows.Win32.Direct2D", "D2D_MATRIX_3X2_F"),
             ("Windows.Win32.SystemServices", "LARGE_INTEGER"),
             ("Windows.Win32.SystemServices", "ULARGE_INTEGER"),
