+++ crates/gen/src/types/struct.rs
@@ -257,6 +257,105 @@ impl Struct {
 
     fn gen_replacement(&self) -> Option<TokenStream> {
         match self.0.full_name() {
+            ("Windows.Win32.SystemServices", "BOOL") => Some(quote! {
+                #[repr(C)]
+                #[allow(non_snake_case)]
+                #[derive(::std::clone::Clone, ::std::marker::Copy, ::std::cmp::PartialEq, ::std::cmp::Eq, ::std::default::Default)]
+                pub struct BOOL(pub i32);
+                impl BOOL {
+                    #[inline]
+                    pub fn as_bool(self) -> bool {
+                        if self.0 == 0 {
+                            false
+                        } else {
+                            true
+                        }
+                    }
+                    #[inline]
+                    pub fn ok(self) -> ::windows::Result<()> {
+                        if self.as_bool() {
+                            Ok(())
+                        } else {
+                            Err(::windows::ErrorCode::from_thread().into())
+                        }
+                    }
+                    #[inline]
+                    pub fn unwrap(self) {
+                        self.ok().unwrap();
+                    }
+                    #[inline]
+                    pub fn expect(self, msg: &str) {
+                        self.ok().expect(msg);
+                    }
+                }
+                impl ::std::fmt::Debug for BOOL {
+                    fn fmt(&self, fmt: &mut ::std::fmt::Formatter<'_>) -> ::std::fmt::Result {
+                        let msg = if self.as_bool() { "true" } else { "false" };
+                        fmt.write_str(msg)
+                    }
+                }
+                unsafe impl ::windows::Abi for BOOL {
+                    type Abi = Self;
+                }
+                impl ::std::convert::From<BOOL> for bool {
+                    fn from(value: BOOL) -> Self {
+                        value.as_bool()
+                    }
+                }
+                
+                impl ::std::convert::From<&BOOL> for bool {
+                    fn from(value: &BOOL) -> Self {
+                        value.as_bool()
+                    }
+                }
+                
+                impl ::std::convert::From<bool> for BOOL {
+                    fn from(value: bool) -> Self {
+                        if value {
+                            BOOL(1)
+                        } else {
+                            BOOL(0)
+                        }
+                    }
+                }
+                
+                impl ::std::convert::From<&bool> for BOOL {
+                    fn from(value: &bool) -> Self {
+                        (*value).into()
+                    }
+                }
+
+
+                
+                impl ::std::cmp::PartialEq<bool> for BOOL {
+                    fn eq(&self, other: &bool) -> bool {
+                        self.as_bool() == *other
+                    }
+                }
+                
+                impl ::std::cmp::PartialEq<BOOL> for bool {
+                    fn eq(&self, other: &BOOL) -> bool {
+                        *self == other.as_bool()
+                    }
+                }
+                
+                impl std::ops::Not for BOOL {
+                    type Output = Self;
+                    fn not(self) -> Self::Output {
+                        if self.as_bool() {
+                            BOOL(0)
+                        } else {
+                            BOOL(1)
+                        }
+                    }
+                }
+                impl<'a> ::windows::IntoParam<'a, BOOL> for bool {
+                    fn into_param(self) -> ::windows::Param<'a, BOOL> {
+                        ::windows::Param::Owned(self.into())
+                    }
+                }
+            }),
+
             ("Windows.Win32.SystemServices", "PWSTR") => Some(quote! {
                 #[repr(C)]
                 #[allow(non_snake_case)]
@@ -296,6 +395,7 @@ impl Struct {
                     }
                 }
             }),
+            // TODO: This can be an extension rather than replacement
             ("Windows.Win32.Automation", "BSTR") => Some(quote! {
                 #[repr(C)]
                 #[allow(non_snake_case)]
