+++ src/lib.rs
@@ -21,7 +21,7 @@ use interfaces::*;
 use runtime::*;
 
 pub use interfaces::{IActivationFactory, IAgileObject, IUnknown, Object};
-pub use result::{Error, ErrorCode, Result, BOOL, FALSE, TRUE};
+pub use result::{Error, ErrorCode, Result};
 pub use runtime::{
     create_instance, factory, initialize_mta, initialize_sta, Array, FactoryCache, Guid, Param,
     RefCount, Waiter,
