+++ src/result/mod.rs
@@ -1,8 +1,6 @@
-mod bool32;
 mod error;
 mod error_code;
 
-pub use bool32::*;
 pub use error::*;
 pub use error_code::*;
